refactor(data-feeds-config-generator): Move CL network name type to `config-types` package
