+++ libs/ts/config-types/src/chainlink-compatibility/types.ts
@@ -1,11 +1,69 @@
 import { Schema as S } from 'effect';
 
-import { networkName, ethereumAddress } from '@blocksense/base-utils/evm';
+import { ethereumAddress } from '@blocksense/base-utils/evm';
+import { keysOf, valuesOf } from '@blocksense/base-utils';
 
-const [_local, ...networkNames] = networkName.literals;
+export const chainlinkNetworkNameToChainId = {
+  'avalanche-fuji-testnet': 'avalanche-fuji',
+  'avalanche-mainnet': 'avalanche-mainnet',
+  'bsc-mainnet': 'bsc-mainnet',
+  'bsc-testnet': 'bsc-testnet',
+  'celo-mainnet': 'celo-mainnet',
+  'celo-testnet-alfajores': 'celo-alfajores',
+  'ethereum-mainnet-andromeda-1': 'metis-andromeda-mainnet',
+  'ethereum-mainnet-arbitrum-1': 'arbitrum-mainnet',
+  'ethereum-mainnet-base-1': 'base-mainnet',
+  'ethereum-mainnet-linea-1': 'linea-mainnet',
+  'ethereum-mainnet-optimism-1': 'optimism-mainnet',
+  'ethereum-mainnet-polygon-zkevm-1': 'polygon-zkevm-mainnet',
+  'ethereum-mainnet-scroll-1': 'scroll-mainnet',
+  'ethereum-mainnet-starknet-1': null,
+  'ethereum-mainnet-zksync-1': 'zksync-mainnet',
+  'ethereum-testnet-sepolia-arbitrum-1': 'arbitrum-sepolia',
+  'ethereum-testnet-sepolia-base-1': 'base-sepolia',
+  'ethereum-testnet-sepolia-optimism-1': 'optimism-sepolia',
+  'ethereum-testnet-sepolia-polygon-zkevm-1': 'polygon-zkevm-cardona',
+  'ethereum-testnet-sepolia-scroll-1': 'scroll-sepolia',
+  'ethereum-testnet-sepolia-starknet-1': null,
+  'ethereum-testnet-sepolia-zksync-1': 'zksync-sepolia',
+  'ethereum-testnet-sepolia': 'ethereum-sepolia',
+  'fantom-mainnet': 'fantom-mainnet',
+  'fantom-testnet': 'fantom-testnet',
+  'kusama-mainnet-moonriver': 'kusama-moonriver',
+  mainnet: 'ethereum-mainnet',
+  'matic-mainnet': 'polygon-mainnet',
+  'polkadot-mainnet-moonbeam': null,
+  'polygon-testnet-amoy': 'polygon-amoy',
+  'solana-devnet': null,
+  'solana-mainnet': null,
+  'xdai-mainnet': 'gnosis-mainnet',
+} as const;
+
+export const chainlinkNetworkName = S.Literal(
+  ...keysOf(chainlinkNetworkNameToChainId),
+);
+
+export type ChainlinkNetworkName = typeof chainlinkNetworkName.Type;
+
+export const parseNetworkFilename = S.decodeUnknownSync(chainlinkNetworkName);
+
+export const chainlinkSupportedNetworkFileName = S.TemplateLiteral(
+  S.Literal('feeds-'),
+  chainlinkNetworkName,
+  S.Literal('.json'),
+);
+
+export type ChainlinkSupportedNetworkFileName =
+  typeof chainlinkSupportedNetworkFileName.Type;
+
+export const isChainlinkSupportedNetworkFileName = S.is(
+  chainlinkSupportedNetworkFileName,
+);
 
 export const ChainlinkAggregatorsSchema = S.Record({
-  key: S.Literal(...networkNames),
+  key: S.Literal(
+    ...valuesOf(chainlinkNetworkNameToChainId).filter(x => x !== null),
+  ),
   value: S.UndefinedOr(ethereumAddress),
 }).annotations({ identifier: 'ChainlinkAggregators' });
